+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.3.2
+
+  - use web request instead bits transfert in powershell toolchain
+  - move check project name toolchain code
+  - move check rule toolchain code
+
 ## 2.3.1
 
   - make doxygen install as optional